+++ .build.ps1
@@ -138,7 +138,7 @@ task Test {
 	$output = Invoke-Build . Tests\.build.ps1 -Result result -Summary | Out-String -Width:200
 	if ($NoTestDiff) {return}
 
-	assert (243 -eq $result.Tasks.Count) $result.Tasks.Count
+	assert (244 -eq $result.Tasks.Count) $result.Tasks.Count
 	assert (60 -eq $result.Errors.Count) $result.Errors.Count
 	assert ($result.Warnings.Count -ge 4)